+++ src/model/FullEmployee.java
@@ -6,7 +6,7 @@ public class FullEmployee extends Employee {
     public FullEmployee() {
     }
 
-    public FullEmployee(int id, String name, int age, String gender, String phone, String email, boolean status, double salary) {
+    public FullEmployee(int id, String name, int age, String gender, String phone, String email, boolean status, int salary) {
         super(id, name, age, gender, phone, email, status, salary);
     }
