+++ src/model/PartEmployee.java
@@ -6,7 +6,7 @@ public class PartEmployee extends Employee {
     public PartEmployee() {
     }
 
-    public PartEmployee(int id, String name, int age, String gender, String phone, String email, boolean status, double salary, int timeWorking) {
+    public PartEmployee(int id, String name, int age, String gender, String phone, String email, boolean status, int salary, int timeWorking) {
         super(id, name, age, gender, phone, email, status, salary);
         this.timeWorking = timeWorking;
     }
